+++ openmc_dagmc_wrapper/Tally.py
@@ -134,128 +134,6 @@ class MeshTallies3D:
             )
 
 
-class MeshTally2D(Tally):
-    """[summary]
-
-    Args:
-        tally_type (str): [description]
-        plane (str): "xy", "xz", "yz"
-        mesh_resolution (list): [description]
-        bounding_box ([type], optional): either a .h5m filename or
-            [point1, point2]. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        tally_type: str,
-        plane: str,
-        bounding_box: List[Tuple[float]],
-        plane_slice_location: Tuple[float, float] = (1, -1),
-        mesh_resolution: Tuple[float, float] = (400, 400),
-    ):
-        self.tally_type = tally_type
-        self.plane = plane
-        self.mesh_resolution = mesh_resolution
-        self.bounding_box = bounding_box
-        self.plane_slice_location = plane_slice_location
-
-        self.create_mesh(bounding_box)
-
-        super().__init__(tally_type)
-        self.name = self.tally_type + "_on_2D_mesh_" + self.plane
-        self.filters.append(openmc.MeshFilter(self.mesh))
-
-    def create_mesh(self, bounding_box):
-        mesh_name = "2D_mesh_" + self.plane
-        mesh = openmc.RegularMesh(name=mesh_name)
-
-        # mesh dimension
-        if self.plane == "xy":
-            mesh.dimension = [
-                self.mesh_resolution[0],
-                self.mesh_resolution[1],
-                1,
-            ]
-            mesh.lower_left = [
-                self.bounding_box[0][0],
-                self.bounding_box[0][1],
-                self.plane_slice_location[1],
-            ]
-            mesh.upper_right = [
-                self.bounding_box[1][0],
-                self.bounding_box[1][1],
-                self.plane_slice_location[0],
-            ]
-
-        elif self.plane == "xz":
-            mesh.dimension = [
-                self.mesh_resolution[0],
-                1,
-                self.mesh_resolution[1],
-            ]
-            mesh.lower_left = [
-                self.bounding_box[0][0],
-                self.plane_slice_location[1],
-                self.bounding_box[0][2],
-            ]
-            mesh.upper_right = [
-                self.bounding_box[1][0],
-                self.plane_slice_location[0],
-                self.bounding_box[1][2],
-            ]
-
-        elif self.plane == "yz":
-            mesh.dimension = [
-                1,
-                self.mesh_resolution[0],
-                self.mesh_resolution[1],
-            ]
-            mesh.lower_left = [
-                self.plane_slice_location[1],
-                self.bounding_box[0][1],
-                self.bounding_box[0][2],
-            ]
-            mesh.upper_right = [
-                self.plane_slice_location[0],
-                self.bounding_box[1][1],
-                self.bounding_box[1][2],
-            ]
-
-        self.mesh = mesh
-
-
-class MeshTallies2D:
-    """[summary]
-
-    Args:
-        tally_types (list): [description]
-        planes (list): list of str with planes
-        meshes_resolutions (list): [description]
-        meshes_corners (list, optional): [description]. Defaults to None.
-        bounding_box ([type], optional): [description]. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        tally_types: str,
-        planes: str,
-        bounding_box: Union[str, List[Tuple[float]]],
-        meshes_resolution: Tuple[float, float] = (400, 400),
-    ):
-        self.tallies = []
-        self.tally_types = tally_types
-        for tally_type in self.tally_types:
-            for plane in planes:
-                self.tallies.append(
-                    MeshTally2D(
-                        tally_type=tally_type,
-                        plane=plane,
-                        mesh_resolution=meshes_resolution,
-                        bounding_box=bounding_box,
-                    )
-                )
-
-
 def compute_filters(tally_type):
     energy_bins_n, dose_coeffs_n = openmc.data.dose_coefficients(
         particle="neutron",
