+++ openmc_dagmc_wrapper/__init__.py
@@ -4,8 +4,6 @@ from .Geometry import Geometry
 from .Materials import Materials
 from .Tally import (
     Tally,
-    MeshTally2D,
-    MeshTallies2D,
     MeshTally3D,
     MeshTallies3D,
 )
