+++ openmc_dagmc_wrapper/tallies/__init__.py
@@ -1,2 +1,3 @@
 from .cell_tally import CellTally, CellTallies
 from .tet_mesh_tallies import TetMeshTally, TetMeshTallies
+from .regular_mesh_tallies import *
