+++ openmc_dagmc_wrapper/tallies/regular_mesh_tallies/__init__.py
@@ -1,0 +1,1 @@
+from .mesh_tally_2D import MeshTally2D, MeshTallies2D
